Fix broken tests
Moved corp enums to separate file to prevent use-before-defined errors
