--- src/Corporation/EmployeePositions.ts
@@ -1,9 +1,0 @@
-export enum EmployeePositions {
-  Operations = "Operations",
-  Engineer = "Engineer",
-  Business = "Business",
-  Management = "Management",
-  RandD = "Research & Development",
-  Training = "Training",
-  Unassigned = "Unassigned",
-}
