+++ src/Corporation/data/Enums.ts
@@ -1,0 +1,27 @@
+export enum IndustryType {
+  Energy = "Energy",
+  Utilities = "Water Utilities",
+  Agriculture = "Agriculture",
+  Fishing = "Fishing",
+  Mining = "Mining",
+  Food = "Food",
+  Tobacco = "Tobacco",
+  Chemical = "Chemical",
+  Pharmaceutical = "Pharmaceutical",
+  Computers = "Computer Hardware",
+  Robotics = "Robotics",
+  Software = "Software",
+  Healthcare = "Healthcare",
+  RealEstate = "Real Estate",
+}
+
+export enum EmployeePositions {
+    Operations = "Operations",
+    Engineer = "Engineer",
+    Business = "Business",
+    Management = "Management",
+    RandD = "Research & Development",
+    Training = "Training",
+    Unassigned = "Unassigned",
+  }
+
